Scrap controller, service 이동
